FIX: Saves training and testing folders
